# server/docs/swagger.yaml
basePath: /
definitions:
  model.Analysis:
    properties:
      claim:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      end_time:
        type: integer
      end_time_fraction:
        type: number
      fact:
        type: string
      id:
        type: integer
      rating:
        $ref: '#/definitions/model.Rating'
        type: object
      rating_id:
        type: integer
      start_time:
        type: integer
      updated_at:
        type: string
      video:
        $ref: '#/definitions/model.Video'
        type: object
      video_id:
        type: integer
    type: object
  model.Rating:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      numeric_value:
        type: integer
      slug:
        type: string
      space_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.Video:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      space_id:
        type: integer
      summary:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      video_type:
        type: string
    type: object
...
  video.paging:
    properties:
      total:
        type: integer
      videos:
        items:
          $ref: '#/definitions/model.Video'
        type: array
    type: object
  video.video:
    properties:
      space_id:
        type: integer
      summary:
        type: string
      title:
        type: string
      url:
        type: string
      video_type:
        type: string
    required:
    - title
    - url
    - video_type
    type: object
  video.videoanalysis:
    properties:
      claim:
        type: string
      end_time:
        type: integer
      end_time_fraction:
        type: number
      fact:
        type: string
      id:
        type: integer
      rating_id:
        type: integer
      start_time:
        type: integer
    required:
    - end_time
    - end_time_fraction
    - rating_id
    type: object
  video.videoanalysisData:
    properties:
      analysis:
        items:
          $ref: '#/definitions/model.Analysis'
        type: array
      video:
        $ref: '#/definitions/model.Video'
        type: object
    type: object
  video.videoanalysisReqData:
    properties:
      analysis:
        items:
          $ref: '#/definitions/video.videoanalysis'
        type: array
      video:
        $ref: '#/definitions/video.video'
        type: object
    required:
    - analysis
    - video
    type: object
...
  videoanalysis.video:
    properties:
      summary:
        type: string
      title:
        type: string
      url:
        type: string
      video_type:
        type: string
    required:
    - title
    - url
    - video_type
    type: object
  videoanalysis.videoAnalysis:
    properties:
      claim:
        type: string
      end_time:
        type: integer
      end_time_fraction:
        type: number
      fact:
        type: string
      rating_id:
        type: integer
      start_time:
        type: integer
    required:
    - end_time
    - end_time_fraction
    - rating_id
    type: object
  videoanalysis.videoAnalysisApiData:
    properties:
      analysis:
        items:
          $ref: '#/definitions/videoanalysis.videoAnalysis'
        type: array
      video:
        $ref: '#/definitions/videoanalysis.video'
        type: object
    required:
    - analysis
    - video
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: VidCheck server API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: VidCheck API
  version: "1.0"
paths:
  /analysis:
    post:
      description: Create videoAnalysis
      operationId: add-video-analysis
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Organisation ID
        in: header
        name: X-Organisation
        required: true
        type: string
      - description: Video Analysis Api Data
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/videoanalysis.videoAnalysisApiData'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Video'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Create videoAnalysis
      tags:
      - VideoAnalysis
  /analysis/video/{video_id}/analysis/{video_analysis_id}:
    delete:
      description: Delete videoAnalysis by id
      operationId: delete-video-analysis-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Organisation ID
        in: header
        name: X-Organisation
        required: true
        type: string
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Video Analysis ID
        in: path
        name: video_analysis_id
        required: true
        type: string
      responses:
        "200": {}
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Delete a videoAnalysis by id
      tags:
      - VideoAnalysis
    get:
      description: Get videosAnalysis by ID
      operationId: get-video-analysis-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Organisation ID
        in: header
        name: X-Organisation
        required: true
        type: string
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Video Analysis ID
        in: path
        name: video_analysis_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Analysis'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Show a videosAnalysis by id
      tags:
      - VideoAnalysis
    put:
      description: Update videosAnalysis by ID
      operationId: update-video-analysis-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Organisation ID
        in: header
        name: X-Organisation
        required: true
        type: string
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Video Analysis ID
        in: path
        name: video_analysis_id
        required: true
        type: string
      - description: Video Analysis Data
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/videoanalysis.videoAnalysis'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Analysis'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Update a videosAnalysis by id
      tags:
      - VideoAnalysis
...
  /videos:
    get:
      description: Get all videos
      operationId: get-all-videos
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: limt per page
        in: query
        name: limit
        type: string
      - description: page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/video.paging'
      summary: Show all videos
      tags:
      - Videos
    post:
      description: Create video
      operationId: add-video
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Video Analysis Api Data
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/video.videoanalysisReqData'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/video.videoanalysisData'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Create video
      tags:
      - Videos
  /videos/{video_id}:
    delete:
      description: Delete video by ID
      operationId: delete-video-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: string
      responses:
        "200": {}
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Delete a video
      tags:
      - Videos
    get:
      description: Get video by ID
      operationId: get-video by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/video.videoanalysisData'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Show a video by id
      tags:
      - Videos
    put:
      description: Update video by ID
      operationId: update-video-by-id
      parameters:
      - description: User ID
        in: header
        name: X-User
        required: true
        type: string
      - description: Space ID
        in: header
        name: X-Space
        required: true
        type: string
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: string
      - description: Video Analysis Api Data
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/video.videoanalysisReqData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Video'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Update a video by id
      tags:
      - Videos
swagger: "2.0"